Move welcome-popup
