+++ src/frontend/static/css/podprofile/podprofile.css
@@ -319,17 +319,30 @@ body.dark-mode .help-button {
 /* Popup Close Button */
 .close-btn {
   position: absolute;
-  top: var(--spacing-sm);
+  top: var(--spacing-md);
   right: var(--spacing-md);
+  width: 40px;
+  height: 40px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   font-size: 1.5rem;
   font-weight: bold;
+  color: var(--text-color-light);
+  background-color: white;
+  border-radius: 50%;
   cursor: pointer;
   color: #555;
   transition: color 0.3s ease;
+  transition: all var(--transition-normal);
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  z-index: 10;
 }
 
 .close-btn:hover {
   color: var(--highlight-color);
+  transform: rotate(90deg);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.15);
 }
 
 /* Popup Titles */
@@ -384,6 +397,119 @@ body.dark-mode .help-button {
 .label-group .help-button:hover + .help-text {
   visibility: visible;
   opacity: 1;
+}
+
+.save-btn {
+  background-color: var(--highlight-color);
+  color: white;
+  box-shadow: var(--button-shadow);
+  position: relative;
+  overflow: hidden;
+}
+
+.save-btn::after {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: -100%;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(
+    90deg,
+    transparent,
+    rgba(255, 255, 255, 0.2),
+    transparent
+  );
+  transition: 0.5s;
+}
+
+.save-btn:hover {
+  background-color: var(--highlight-hover);
+  transform: translateY(-2px);
+  box-shadow: 0 6px 10px -3px rgba(0, 0, 0, 0.2);
+}
+
+.save-btn:hover::after {
+  left: 100%;
+}
+
+.popup-content h2 {
+  margin-bottom: var(--spacing-lg);
+  color: rgba(0, 0, 0, 0.8);
+  text-align: center;
+  font-size: 1.8rem;
+  font-weight: 700;
+  position: relative;
+  padding-bottom: 15px;
+}
+
+.popup-content h2::after {
+  content: "";
+  position: absolute;
+  bottom: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 60px;
+  height: 3px;
+  background-color: var(--highlight-color);
+  border-radius: 3px;
+}
+
+.popup-body {
+  color: var(--text-color-light);
+}
+
+.popup-body p {
+  margin-bottom: var(--spacing-md);
+  line-height: 1.6;
+}
+
+.popup-body .btn {
+  margin-top: var(--spacing-lg);
+  display: block;
+  width: 100%;
+  padding: var(--spacing-md);
+}
+
+.popup {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  display: none;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000; /* High z-index but below header */
+  backdrop-filter: blur(5px);
+  transition: all 0.4s cubic-bezier(0.175, 0.885, 0.32, 1.275);
+  opacity: 0;
+  visibility: hidden;
+}
+
+.popup[style*="display: flex"] {
+  opacity: 1;
+  visibility: visible;
+}
+
+.popup-content {
+  background-color: white;
+  border-radius: var(--radius-large);
+  box-shadow: 0 25px 50px -12px rgba(0, 0, 0, 0.25);
+  padding: var(--spacing-xl);
+  width: 90%;
+  max-width: 600px;
+  position: relative;
+  transform: scale(0.95);
+  transition: transform 0.5s cubic-bezier(0.175, 0.885, 0.32, 1.275);
+  overflow-x: hidden;
+  max-height: 85vh;
+  overflow-y: auto;
+}
+
+.popup[style*="display: flex"] .popup-content {
+  transform: scale(1);
 }
 
 /* NEW UNIFIED PODCAST CONTAINER */
@@ -930,6 +1056,12 @@ body.dark-mode .error-container {
 
   .episodes-list {
     margin-top: 0; /* Remove unnecessary margin */
+  }
+}
+
+@media (max-width: 576px) {
+  .popup-content {
+    padding: var(--spacing-lg);
   }
 }
