+++ src/frontend/templates/podprofile/podprofile.html
@@ -118,6 +118,33 @@
         <!-- Podcast info will be inserted here by JavaScript -->
       </div>
     </main>
+
+    <!-- Welcome Popup -->
+  <div id="welcome-popup" class="popup">
+    <div class="popup-content">
+      <span id="close-welcome-popup" class="close-btn">&times;</span>
+      <h2>Welcome to PodManager.ai!</h2>
+      <div class="popup-body">
+        <p>🎉 <strong>Congratulations!</strong></p>
+        <p>
+          Your spot in the Beta Tester queue is secured! Get ready for your
+          exclusive sneak peek...
+        </p>
+        <p>
+          💰 To get you started, we've added <strong>3,500 credits</strong> to
+          your account!
+        </p>
+        <p>
+          PodManager.ai will revolutionize your podcast production with
+          AI-driven workflow automation, seamless guest management, and faster
+          episode production - so you can focus on creating amazing content!
+        </p>
+        <p>✨ Check your email now for more exciting features!</p>
+        <button id="get-started-btn" class="btn save-btn">Get Started</button>
+      </div>
+    </div>
+  </div>
+  
     <script
       type="module"
       src="{{ url_for('static', filename='js/podprofile/podprofile.js') }}"
